Frame name changed
