+++ tid_project/src/App.js
@@ -5,13 +5,13 @@ import { theme } from './styles/theme';
 import { GlobalStyles } from './styles/GlobalStyles';
 import DashboardLayout from './components/layout/DashboardLayout';
 import Frame1 from './pages/Frame1/Frame1';
-import Frame3 from './pages/Frame3/Frame3';
+import ContentPage from './pages/ContentPage/ContentPage';
 import Threads from './pages/Threads/Threads';
 import Login from './pages/Login/Login';
 import Settings from './pages/Settings/Settings';
 import apiClient from './api/client';
 import CreateContentPage from './pages/CreateContent/CreateContentPage';
-import SingleContentPage from './pages/ContentPages/SingleContentPage';
+import SingleContentPage from './pages/ContentPage/SingleContentPage';
 import ProfileList from './pages/ProfileList/ProfileList';
 
 const checkUser = async () => {
@@ -41,7 +41,7 @@ const ProtectedRoute = ({ children }) => {
   }, []);
 
   if (isAuthenticated === null) {
-    return <div>Loading...</div>; 
+    return <div>Loading...</div>;
   }
 
   if (!isAuthenticated) {
@@ -63,7 +63,7 @@ const PublicRoute = ({ children }) => {
   }, []);
 
   if (isAuthenticated === null) {
-    return <div>Loading...</div>; 
+    return <div>Loading...</div>;
   }
 
   if (isAuthenticated) {
@@ -80,16 +80,16 @@ function App() {
       <div>
         <BrowserRouter>
           <Routes>
-            <Route 
-              path="/login" 
+            <Route
+              path="/login"
               element={
                 <PublicRoute>
                   <Login />
                 </PublicRoute>
-              } 
+              }
             />
-            <Route 
-              path="/" 
+            <Route
+              path="/"
               element={
                 <ProtectedRoute>
                   <DashboardLayout />
@@ -99,11 +99,11 @@ function App() {
               <Route index element={<Navigate to="/frame1" replace />} />
               <Route path="frame1" element={<Frame1 />} />
               <Route path="profilelist" element={<ProfileList />} />
-              <Route path="frame3" element={<Frame3 />} />
+              <Route path="contentpage" element={<ContentPage />} />
               <Route path="Threads" element={<Threads />} />
               <Route path="settings" element={<Settings />} />
               <Route path="/create-content" element={<CreateContentPage />} />
-              <Route path="/content/:type/:id" element={<SingleContentPage/>} />
+              <Route path="/content/:type/:id" element={<SingleContentPage />} />
             </Route>
           </Routes>
         </BrowserRouter>
