+++ tid_project/src/pages/ContentPage/Card.js
